backend/pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "artemowkin", email = "artyom.groshev2017@gmail.com"},
]
dependencies = [
    "fastapi>=0.98.0",
    "sqlalchemy>=2.0.17",
    "uvicorn>=0.22.0",
    "psycopg>=3.1.9",
    "passlib>=1.7.4",
    "redis>=4.6.0",
    "email-validator>=2.0.0.post2",
    "python-jose>=3.3.0",
    "pycrypto>=2.6.1",
    "rsa>=4.9",
    "python-dotenv>=1.0.0",
    "ipython>=8.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
